+++ service/src/utils/mod.rs
@@ -1,5 +1,6 @@
 pub mod github;
 pub mod wget;
 pub mod download;
+pub mod zip_extract;